// src/generated-api/models/ReportFindAllResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BMI-LAB Web API
 * BMI-LAB을 관리하기 위한 웹 서비스
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ReportSummary } from './ReportSummary';
import {
  ReportSummaryFromJSON,
  ReportSummaryFromJSONTyped,
  ReportSummaryToJSON,
  ReportSummaryToJSONTyped,
} from './ReportSummary';

/**
 *
 * @export
 * @interface ReportFindAllResponse
 */
export interface ReportFindAllResponse {
  /**
   *
   * @type {Array<ReportSummary>}
   * @memberof ReportFindAllResponse
   */
  reports?: Array<ReportSummary>;
}

/**
 * Check if a given object implements the ReportFindAllResponse interface.
 */
export function instanceOfReportFindAllResponse(
  value: object,
): value is ReportFindAllResponse {
  return true;
}

export function ReportFindAllResponseFromJSON(
  json: any,
): ReportFindAllResponse {
  return ReportFindAllResponseFromJSONTyped(json, false);
}

export function ReportFindAllResponseFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): ReportFindAllResponse {
  if (json == null) {
    return json;
  }
  return {
    reports:
      json['reports'] == null
        ? undefined
        : (json['reports'] as Array<any>).map(ReportSummaryFromJSON),
  };
}

export function ReportFindAllResponseToJSON(json: any): ReportFindAllResponse {
  return ReportFindAllResponseToJSONTyped(json, false);
}

export function ReportFindAllResponseToJSONTyped(
  value?: ReportFindAllResponse | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    reports:
      value['reports'] == null
        ? undefined
        : (value['reports'] as Array<any>).map(ReportSummaryToJSON),
  };
}