// src/generated-api/models/SearchProjectItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BMI-LAB Web API
 * BMI-LAB을 관리하기 위한 웹 서비스
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SearchProjectItem
 */
export interface SearchProjectItem {
    /**
     * 
     * @type {number}
     * @memberof SearchProjectItem
     */
    projectId?: number;
    /**
     * 
     * @type {string}
     * @memberof SearchProjectItem
     */
    title?: string;
}

/**
 * Check if a given object implements the SearchProjectItem interface.
 */
export function instanceOfSearchProjectItem(value: object): value is SearchProjectItem {
    return true;
}

export function SearchProjectItemFromJSON(json: any): SearchProjectItem {
    return SearchProjectItemFromJSONTyped(json, false);
}

export function SearchProjectItemFromJSONTyped(json: any, ignoreDiscriminator: boolean): SearchProjectItem {
    if (json == null) {
        return json;
    }
    return {
        
        'projectId': json['projectId'] == null ? undefined : json['projectId'],
        'title': json['title'] == null ? undefined : json['title'],
    };
}

export function SearchProjectItemToJSON(json: any): SearchProjectItem {
    return SearchProjectItemToJSONTyped(json, false);
}

export function SearchProjectItemToJSONTyped(value?: SearchProjectItem | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'projectId': value['projectId'],
        'title': value['title'],
    };
}